+++ src/test/java/org/example/Room/RoomTest.java
@@ -1,11 +1,9 @@
-package org.example;
+package org.example.Room;
 
 import org.example.Item.Potion.Potion;
 import org.example.Item.Potion.Heal.SmallHealingPotion;
 import org.example.Item.Weapon.Sword;
 import org.example.Item.Weapon.Weapon;
-import org.example.Room.Direction;
-import org.example.Room.Room;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
